requirements.txt
# Install torch:
# pip install torch==1.9.0+cu111 torchvision==0.10.0+cu111 torchaudio==0.9.0 -f https://download.pytorch.org/whl/torch_stable.html

h5py == 3.1.0
# transformers 4.11.0 above requires "huggingface-hub>=0.0.17"
transformers == 4.10.3 
huggingface-hub == 0.0.12
numpy == 1.19.5
Pillow == 8.3.1
#torchvision
tqdm == 4.61.2
# Newly added for the Attention visualization: cv2, matplot
opencv-python == 4.7.0.68
matplotlib == 3.2.2